wasm: resample event-sound channels to output rate in audio_mix
audio_mix advanced pcm_position by 1 per output frame regardless of
the source's actual sample_rate. With the hardware output fixed at
44100 Hz, a 22050 Hz Snake sound effect played 2× faster and 2× higher
in pitch (classic chipmunk). The out_rate parameter was passed in but
never consulted inside the channel loop.

Promote pcm_position from size_t to double so it can hold a fractional
read index, then advance by rate_ratio = ch->sample_rate / out_rate
per output frame (nearest-neighbor resampling — adequate for game SFX
and matches the existing one-tap-per-frame structure).

Streaming-sound path (ctx->stream.read_pos++) has the same bug but
isn't exercised by any current test SWF — follow-up.

+++ SWFModernRuntime/include/audio/audio.h
@@ -44,7 +44,7 @@ typedef struct SoundChannel {
 	const SoundAsset* asset;
 	float* pcm_data;      // Decoded PCM (float, interleaved)
 	size_t pcm_samples;   // Total decoded samples (per channel)
-	size_t pcm_position;  // Current playback position
+	double pcm_position;  // Current playback position (fractional, resampled)
 	int channels;         // 1=mono, 2=stereo
 	int sample_rate;      // Actual sample rate in Hz
 	u32 loop_count;       // 0 = no loop, >0 = remaining loops
